requirements.txt
astropy
numpy
Pillow
